REGRESSION (276531@main): Content flash when expanding image viewer on eBay page https://bugs.webkit.org/show_bug.cgi?id=275739 rdar://128178476
Reviewed by NOBODY (OOPS!).

It makes no sense that `affectsRenderedSubtree()` would return `true` if we stored a `RenderStyle` for `display: none`,
so we remove that case. However this change regressed a couple of WTP tests:

    - imported/w3c/web-platform-tests/css/css-transitions/starting-style-rule-basic.html
    - imported/w3c/web-platform-tests/css/css-transitions/starting-style-rule-pseudo-elements.html

The issue there is that the stored `display: none` style would only be changed upon `display` changing to `none`, under
`RenderTreeUpdater::updateElementRenderer()`, but any further style change while `display` remained `none` would not
update that stored `display: none` style since that method was not called.

We address this by returning the `ElementUpdate` when the resolved style has `display: none` and we already had a stored
`display: none` style.

* Source/WebCore/style/StyleTreeResolver.cpp:
(WebCore::Style::affectsRenderedSubtree):
(WebCore::Style::TreeResolver::resolveElement):