[css-animations] animation-name assigned by ::part() cannot find animation defined in shadow tree https://bugs.webkit.org/show_bug.cgi?id=296048 rdar://156484228
Reviewed by NOBODY (OOPS!).

When looking for a matching animation name for a given element, we would always use the style scope
returned by `Style::Scope::forOrdinal()` providing the `Style::ScopedName` specified on the `Animation`
object we were processing. In the case where the `animation-name` property is specified in the
same scope as the `@keyframes` at-rule with that same name, this works fine. However, in the case
where the `@keyframes` at-rule is defined within a different scope, a shadow root for instance,
then we would fail.

To address this, we add a new `Style::Scope::forScopedAnimationName()` method which will use the
provided `Element` and `Style::ScopedName` pair to find the most relevant style scope where this
animation name is defined, using the element's scope if hosted within a shadow root if another
scope couldn't be found.

We also add two new WPT tests to check that we have the correct behavior in both of the cases
described above.

* LayoutTests/imported/w3c/web-platform-tests/css/css-animations/animation-name-in-shadow-part-expected.html: Added.
* LayoutTests/imported/w3c/web-platform-tests/css/css-animations/animation-name-in-shadow-part-inner-match-expected.html: Added.
* LayoutTests/imported/w3c/web-platform-tests/css/css-animations/animation-name-in-shadow-part-inner-match.html: Added.
* LayoutTests/imported/w3c/web-platform-tests/css/css-animations/animation-name-in-shadow-part-outer-match-expected.html: Added.
* LayoutTests/imported/w3c/web-platform-tests/css/css-animations/animation-name-in-shadow-part-outer-match.html: Added.
* LayoutTests/imported/w3c/web-platform-tests/css/css-animations/animation-name-in-shadow-part-ref.html: Added.
* LayoutTests/imported/w3c/web-platform-tests/css/css-animations/animation-name-in-shadow-part.html: Added.
* Source/WebCore/animation/KeyframeEffect.cpp:
(WebCore::KeyframeEffect::computeCSSAnimationBlendingKeyframes):
* Source/WebCore/style/StyleScope.cpp:
(WebCore::Style::Scope::forScopedAnimationName):
* Source/WebCore/style/StyleScope.h:
* Source/WebCore/style/Styleable.cpp:
(WebCore::keyframesRuleExistsForAnimation):
(WebCore::Styleable::updateCSSAnimations const):