[threaded-animations] fully accelerated scroll-driven animations should not update styles while scrolling https://bugs.webkit.org/show_bug.cgi?id=302156 rdar://164253776
Reviewed by NOBODY (OOPS!).

When a scroll-driven animation only affects properties that can be animated in the remote layer tree,
we must make sure not to update styles on the main thread in the web process as scrolling occurs.

To check that is indeed the case, we add a counter that tracks the number of times `KeyframeEffect::apply()`
is called and expose it through the `internals` testing object.

Test: webanimations/threaded-animations/scroll-animation-style-updates-while-scrolling.html

* LayoutTests/webanimations/threaded-animations/scroll-animation-style-updates-while-scrolling-expected.txt: Added.
* LayoutTests/webanimations/threaded-animations/scroll-animation-style-updates-while-scrolling.html: Added.
* Source/WebCore/animation/KeyframeEffect.cpp:
(WebCore::KeyframeEffect::apply):
* Source/WebCore/animation/KeyframeEffect.h:
* Source/WebCore/animation/WebAnimation.h:
* Source/WebCore/testing/Internals.cpp:
(WebCore::Internals::numberOfAnimationStyleUpdates const):
* Source/WebCore/testing/Internals.h:
* Source/WebCore/testing/Internals.idl: